pyproject.toml
[tool.poetry]
name = "vk-user-graph-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
py2neo = "^2021.2.4"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"
pydantic-settings = "^2.7.0"
pytest = "^8.3.4"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
